fix namespaces, move files:new:{dir/symlink} to subcommands
