+++ src/App/Commands/Files/SubCommands/FilesNewDirCommand.php
@@ -1,10 +1,10 @@
 <?php
 
 
-namespace Console\App\Commands\Files;
+namespace Lio\App\Commands\Files;
 
 
-use Console\App\Commands\Command;
+use Lio\App\Commands\Command;
 use Exception;
 use GuzzleHttp\Exception\BadResponseException;
 use GuzzleHttp\Exception\GuzzleException;
